pyproject.toml
[project]
name = "ad-employee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django~=5.2",
    "django-cors-headers>=4.9.0",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.0.2",
    "ldap3>=2.9.1",
    "mssql-django>=1.6",
    "pyodbc>=5.3.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
